Removing manual sleep timeout as redis client retries automatically.
The Redis client automatically retries to connect to the server on
connection issues. Rather than waiting for a fixed time period before
connecting, we can leave this logic to the redis client. Looking at the
error thrown, we print out a message informing the user we are in the
waiting for networking state.